socialnetworkapi/requirements.txt
Django>=5.1.7
djangorestframework
djangorestframework-simplejwt
drf-yasg
psycopg2-binary